server/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
openpyxl==3.0.10
google-auth==2.17.3
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0
google-api-python-client==2.84.0
requests==2.28.1
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==1.0.0